Expand swc blueprint reading/wrinting
